api/requirements.txt
flask
flask-sqlalchemy
pymysql
flask-migrate
python-dotenv
requests
zenora
flask-cors
python-jose[cryptography]